Move Athenz to config-provisioning
